failure-prediction-module/requirements.txt
blosc
pandas
scikit-learn==1.4.1.post1
python-dotenv
pymongo